@@ -5,18 +5,30 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=7.0.0",
-        "fideloper/proxy": "~3.3",
-        "laravel/framework": "5.5.*",
+        "php": ">=7.1.3",
+        "camroncade/timezone": "0.1",
+        "doctrine/dbal": "^2.5",
+        "fideloper/proxy": "~4.0",
+        "fzaninotto/faker": "^1.8",
+        "laravel/framework": "5.8.*",
+        "laravel/socialite": "^3.3",
         "laravel/tinker": "~1.0",
+        "mookofe/tail": "^1.0",
+        "paragonie/constant_time_encoding": "^2.0",
+        "php-amqplib/php-amqplib": "^2.10",
+        "pragmarx/google2fa": "^0.7.1",
+        "predis/predis": "^1.1",
         "proengsoft/laravel-jsvalidation": "^2.4",
-        "yajra/laravel-datatables-oracle": "^8.13"
+        "sudiptpa/guid": "^1.0",
+        "sudiptpa/ipstack": "^1.3",
+        "vladimir-yuldashev/laravel-queue-rabbitmq": "^8.3",
+        "webpatser/laravel-uuid": "3.0",
+        "yajra/laravel-datatables-oracle": "9.*"
     },
     "require-dev": {
         "filp/whoops": "~2.0",
-        "fzaninotto/faker": "~1.4",
         "mockery/mockery": "~1.0",
-        "phpunit/phpunit": "~6.0",
+        "phpunit/phpunit": "~7.0",
         "symfony/thanks": "^1.0"
     },
     "autoload": {
