@@ -6,34 +6,37 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Validator;
+use Session;
 use DataTables;
+use DB;
+use Auth;
 use App\Student;
+use App\User;
 
 class TestController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+    
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \DataTables
      */
-    public function index()
-    {
-        //
-    }
-
     public function studentsData()
     {
-        $students = Student::all();
+        // Get ALL Students
+        DB::statement(DB::raw('set @rownum=0'));
+        $students = Student::get(['students.*', 
+                    DB::raw('@rownum  := @rownum  + 1 AS rownum')]);
+        // return DataTables with modified columns
         return DataTables::of($students)
         ->addColumn('names',function($student)
         {
-            return 
-            '<strong>' .$student->name. '</strong>';
-        })
-        ->addColumn('age',function($student)
-        {
-            return 
-            $student->age;
+            return '<strong>' .$student->name. '</strong>';
         })
         ->addColumn('email',function($student)
         {
@@ -43,22 +46,76 @@ class TestController extends Controller
         {
             return $student->father_name;
         })
+        ->addColumn('file',function($student)
+        {
+            return '<a class="btn btn-primary" target="_blank" href="' .route('viewfile',$student->id). '">View</a>';
+        })
+        // generating UPDATE/DELETE button
         ->addColumn('action',function($student)
         {
-            return '<form action="' .route('viewfile',$student->id). '"><input type="submit" value="View" class="btn btn-primary"></form>';
+            return '<a href="" data-toggle="modal" data-target="#updatemodal" data-id="' .encrypt($student->id). '" data-name="'.$student->name.'" data-email="'.$student->email.'" data-father_name="'.$student->father_name.'" data-address="'.$student->address.'" data-date_of_birth="'.$student->date_of_birth.'" class="btnupdate btn btn-warning">Update</a>
+            <a href="" data-toggle="modal" data-target="#deletemodal" data-id="' .encrypt($student->id). '" style="margin-left:5px" class="btndelete btn btn-danger">Delete</a>';
         })
-        ->rawColumns(['names', 'email', 'action'])
+        // render HTML for following columns
+        ->rawColumns(['names', 'email', 'file', 'action'])
         ->make(true);
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Validate Form Request
+     * Encrypt All Fields And Return
      *
+     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function takeRequest(Request $request)
     {
-        //
+        $error_array = [];
+        $success_output = '';
+        $response = [];
+
+        // Validating Data Before Encryption
+        $validation = Validator::make($request->all(),[
+            'email'=>['required','unique:students'],
+            'name'=>['required','regex:/^[a-zA-Z ]*$/'],
+            'address'=>'required',
+            'father_name'=>'required',
+            'date_of_birth'=>'required',
+            'file_path'=>['required','mimes:pdf'],
+        ]);
+        
+        // collect validation errors
+        if ($validation->fails())
+        {
+            foreach($validation->messages()->getMessages() as $field_name => $messages)
+            {
+                $error_array[] = $messages;
+            }
+        }
+        // encrypting request with success output
+        else
+        {
+            $response = [
+                'name'          =>  encrypt($request->name),
+                'email'         =>  encrypt($request->email),
+                'address'       =>  encrypt($request->address),
+                'father_name'   =>  encrypt($request->father_name),
+                'date_of_birth' =>  encrypt($request->date_of_birth),
+                'file_path'          =>  $request->file_path,
+            ];
+            $success_output = 'success';
+        }
+        
+        // output array return with JSON encoded
+        $output = [];
+        $output = [
+            'error'=>$error_array,
+            'success'=>$success_output,
+            'datas'=>$response,
+        ];
+        echo json_encode($output);
+        exit();
+        
     }
 
     /**
@@ -69,17 +126,32 @@ class TestController extends Controller
      */
     public function store(Request $request)
     {
-        // return $request->all();
-        $error_array = array();
+        $error_array = [] ;
+        $error_output = '';
         $success_output = '';
-        $validation = Validator::make($request->all(),[
+        $student = new Student();
+
+        // decrypting values
+        $values = array( 
+            "name"=> decrypt($request->name),
+            "email"=> decrypt($request->email),
+            "address"=> decrypt($request->address),
+            "date_of_birth"=> decrypt($request->date_of_birth),
+            "father_name"=> decrypt($request->father_name),
+            "file_path"=>  $request->file_path
+        );
+
+        // passing through validation
+        $validation = Validator::make($values,[
             'email'=>['required','unique:students'],
             'name'=>['required','regex:/^[a-zA-Z ]*$/'],
             'address'=>'required',
             'father_name'=>'required',
             'date_of_birth'=>'required',
-            'file'=>['required','mimes:pdf'],
+            'file_path'=>['required','mimes:pdf'],
         ]);
+
+        // collecting vlidation errors if exists
         if ($validation->fails())
         {
             foreach($validation->messages()->getMessages() as $field_name => $messages)
@@ -87,67 +159,59 @@ class TestController extends Controller
                 $error_array[] = $messages;
             }
         }
+        // saving into database after validation
         else
         {
-            $student = new Student();
-            $student->name = $request->name;
-            $student->email = $request->email;
-            $student->address = $request->address;
-            $student->father_name = $request->father_name;
-            $student->date_of_birth = $request->date_of_birth;
-            $student->age = \Carbon\Carbon::parse($request->date_of_birth)->age;
-            $student->file = $request->file;
-            $student->save();
-            $id = $student->id;
-            if ($request->hasFile('file'))
+            $student->name          = $values['name'];
+            $student->email         = $values['email'];
+            $student->address       = $values['address'];
+            $student->father_name   = $values['father_name'];
+            $student->date_of_birth = $values['date_of_birth'];
+            $student->age           =  Student::age($values['date_of_birth']);
+            // Checking File Exists
+            if ($request->hasFile('file_path'))
             {
-                // return "File";
-                $file = $request->file('file');
+                $file = $request->file('file_path');
                 $extension = $file->getClientOriginalExtension();
                 Storage::disk('public')->put($file->getFilename().'.'.$extension,  File::get($file));
-                $path = $request->file('file')->getRealPath();
-                $student = Student::find($id);
-                $student->file = $file->getFilename();
-                $student->save();
+                $path = $request->file('file_path')->getRealPath();
+                $student->file_path = $file->getFilename();
             }
-            $success_output =  "sass";
         }
-        $output = array(
-            'error'     =>  $error_array,
+        
+        $student->save();
+        $id = 0;
+        $id = $student->id;
+        if ($id > 0) {
+            $success_output =  "success";
+        } else {
+            $error_output = 'error';
+        }
+        // output array return with JSON encoded
+        $output = [
+            'validate'  =>  $error_array,
+            'error'     =>  $error_output,
             'success'   =>  $success_output
-        );
+        ];
         echo json_encode($output);
+        exit();
     }
 
+    /**
+     * View File Of Specific Student
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
     public function viewfile($id)
     {
+        // Find Student with the ID
         $student = Student::where('id',$id)->first();
-        $path = $student->file;
+        $path = $student->file_path;
+        // Return File Response
         return response()->file(storage_path('app/public/'.$path).'.pdf', [
             'Content-Type' => 'application/pdf'
         ]);
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
     }
 
     /**
@@ -159,7 +223,57 @@ class TestController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $error_array = [];
+        $success_output = '';
+
+        // validating data before updating
+        $validation = Validator::make($request->all(),[
+            'email'=>['required','unique:students,email,'.decrypt($request->id)],
+            'name'=>['required','regex:/^[a-zA-Z ]*$/'],
+            'address'=>'required',
+            'father_name'=>'required',
+            'date_of_birth'=>'required',
+            'file_path'=>['required','mimes:pdf'],
+        ]);
+        // collecting validation errors
+        if ($validation->fails())
+        {
+            foreach($validation->messages()->getMessages() as $field_name => $messages)
+            {
+                $error_array[] = $messages;
+            }
+        }
+        else
+        {
+            // checking if file exists
+            if ($request->hasFile('file_path'))
+            {
+                $file = $request->file('file_path');
+                $extension = $file->getClientOriginalExtension();
+                Storage::disk('public')->put($file->getFilename().'.'.$extension,  File::get($file));
+                $path = $request->file('file_path')->getRealPath();
+                $file_path = $file->getFilename();
+            }
+            $student = Student::findOrFail(decrypt($id));
+            // delete previous file
+            unlink(storage_path("app\public\\".$student->file_path.".pdf"));
+            // update student
+            $student->update($request->except(['token','id']));
+            $student->update([
+                'file_path'=>$file_path,
+                'age' => Student::age($request->date_of_birth)
+            ]);
+            $success_output = 'success';
+            
+        }
+        // output array return in JSON encoded form
+        $output = [];
+        $output = [
+            'error'=>$error_array,
+            'success'=>$success_output
+        ];
+        echo json_encode($output);
+        exit();
     }
 
     /**
@@ -170,6 +284,43 @@ class TestController extends Controller
      */
     public function destroy($id)
     {
-        //
+        // find student with given ID
+        $student = Student::findOrFail(decrypt($id));
+        // delete their file
+        unlink(storage_path("app\public\\".$student->file_path.".pdf"));
+        // record delete
+        $student->delete();
+        return json_encode('success');
     }
+
+    /**
+    *Display Users Index Page
+    *
+    *
+    * @return view
+    */
+    public function usersindex()
+    {
+        $user = User::findOrFail(Auth::id());
+        Session::push('values', [
+                'id'        =>  encrypt($user->id),
+                'name'      =>  encrypt($user->name),
+                'email'     =>  encrypt($user->email),
+                'protected' =>  encrypt($user->is_auth),
+            ]);
+        // return Session::all();
+        return view('user.profile');
+    }
+
+    /**
+    *Update User
+    *
+    *
+    * @return redirect back
+    */
+    public function edituser(Request $request)
+    {
+        
+    }
+
 }
