@@ -110,7 +110,7 @@ return [
 
         'default' => [
             'host' => env('REDIS_HOST', '127.0.0.1'),
-            'password' => env('REDIS_PASSWORD', null),
+            'password' => env('REDIS_PASSWORD', 'mypass'),
             'port' => env('REDIS_PORT', 6379),
             'database' => 0,
         ],
